Create unified provider model
Move initialization logic of collectors and observers into
their own providers. This makes the code a bit easier to
reason about.
