+++ src/providers/observers/mattermost/validation.rs
@@ -33,11 +33,11 @@ impl Validate for MattermostCredentials {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::config::Configuration;
+    use crate::providers::DuckProviderCollection;
 
     #[test]
-    #[should_panic(expected = "Mattermost observer have no ID.")]
+    #[should_panic(expected = "The id \\'\\' is invalid.")]
     fn should_return_error_if_mattermost_id_is_empty() {
         let config = Configuration::from_json(
             r#"
@@ -52,14 +52,16 @@ mod tests {
                                     "url": "https://mattermost.example.com"
                                 }
                             }
-                        }             
+                        }
                     }
                 ]
             }
         "#,
         )
         .unwrap();
-        config.validate().unwrap();
+
+        let collection = DuckProviderCollection::new();
+        collection.get_observers(&config).unwrap();
     }
 
     #[test]
@@ -79,14 +81,16 @@ mod tests {
                                     "url": "https://mattermost.example.com"
                                 }
                             }
-                        }             
+                        }
                     }
                 ]
             }
         "#,
         )
         .unwrap();
-        config.validate().unwrap();
+
+        let collection = DuckProviderCollection::new();
+        collection.get_observers(&config).unwrap();
     }
 
     #[test]
@@ -105,13 +109,15 @@ mod tests {
                                     "url": ""
                                 }
                             }
-                        }             
+                        }
                     }
                 ]
             }
         "#,
         )
         .unwrap();
-        config.validate().unwrap();
+
+        let collection = DuckProviderCollection::new();
+        collection.get_observers(&config).unwrap();
     }
 }
