+++ src/providers/observers/slack/mod.rs
@@ -1,15 +1,17 @@
 use std::collections::HashSet;
 use std::iter::FromIterator;
 
-use reqwest::header::{ACCEPT, CONTENT_TYPE};
-use reqwest::{Client, ClientBuilder};
-
-use log::{error, info};
+use log::info;
 
 use crate::builds::{Build, BuildStatus};
-use crate::config::{SlackConfiguration, SlackCredentials};
-use crate::observers::{Observation, Observer, ObserverInfo};
+use crate::config::SlackConfiguration;
+use crate::providers::observers::{Observation, Observer, ObserverInfo};
 use crate::utils::DuckResult;
+
+use self::client::SlackClient;
+
+mod client;
+mod validation;
 
 pub struct SlackObserver {
     client: SlackClient,
@@ -73,50 +75,5 @@ fn get_message_icon(build: &Build) -> &str {
         BuildStatus::Failed => ":heavy_multiplication_x:",
         BuildStatus::Running => ":shipit:",
         BuildStatus::Unknown => ":question:",
-    }
-}
-
-struct SlackClient {
-    client: Client,
-    credentials: SlackCredentials,
-}
-
-impl SlackCredentials {
-    pub fn get_url(&self) -> &str {
-        match self {
-            SlackCredentials::Webhook { url } => &url[..],
-        }
-    }
-}
-
-impl SlackClient {
-    pub fn new(config: &SlackConfiguration) -> Self {
-        SlackClient {
-            client: ClientBuilder::new().build().unwrap(),
-            credentials: config.credentials.clone(),
-        }
-    }
-
-    pub fn send(&self, message: &str, icon: &str) -> DuckResult<()> {
-        let response = self
-            .client
-            .put(self.credentials.get_url())
-            .header(CONTENT_TYPE, "application/json")
-            .header(ACCEPT, "application/json")
-            .body(
-                json!({
-                    "username": "Duck",
-                    "icon_emoji": icon,
-                    "text": message
-                })
-                .to_string(),
-            )
-            .send()?;
-
-        if !response.status().is_success() {
-            error!("Could not send Slack message ({})!", response.status());
-        }
-
-        Ok(())
     }
 }
