+++ src/providers/observers/slack/validation.rs
@@ -21,11 +21,11 @@ impl Validate for SlackConfiguration {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::config::Configuration;
+    use crate::providers::DuckProviderCollection;
 
     #[test]
-    #[should_panic(expected = "Slack observer have no ID.")]
+    #[should_panic(expected = "The id \\'\\' is invalid.")]
     fn should_return_error_if_slack_id_is_empty() {
         let config = Configuration::from_json(
             r#"
@@ -40,14 +40,16 @@ mod tests {
                                     "url": "https://slack.com/MY-WEBHOOK-URL"
                                 }
                             }
-                        }             
+                        }
                     }
                 ]
             }
         "#,
         )
         .unwrap();
-        config.validate().unwrap();
+
+        let collection = DuckProviderCollection::new();
+        collection.get_observers(&config).unwrap();
     }
 
     #[test]
@@ -66,13 +68,15 @@ mod tests {
                                     "url": ""
                                 }
                             }
-                        }             
+                        }
                     }
                 ]
             }
         "#,
         )
         .unwrap();
-        config.validate().unwrap();
+
+        let collection = DuckProviderCollection::new();
+        collection.get_observers(&config).unwrap();
     }
 }
